restruct fused strategies and fix test
