+++ python/deep_ep/deep_ep/strategies/__init__.py
@@ -9,13 +9,17 @@ separated by mode:
 
 from ..ep_strategy import (
     EPCommStrategy,
+    FusedEPStrategy,
     LowLatencyEPCommStrategy,
     NormalEPCommStrategy,
+    get_fused_strategy,
     get_low_latency_strategy,
     get_normal_strategy,
+    register_fused_strategy,
     register_low_latency_strategy,
     register_normal_strategy,
 )
+from .fused_strategy import DeepEPFusedStrategy, MegaMoeFusedStrategy
 from .low_latency_strategy import (
     DefaultLowLatencyCommStrategy,
     OpsLowLatencyCommStrategy,
@@ -27,15 +31,21 @@ __all__ = [
     "EPCommStrategy",
     "NormalEPCommStrategy",
     "LowLatencyEPCommStrategy",
+    "FusedEPStrategy",
     # Registry functions
     "register_normal_strategy",
     "register_low_latency_strategy",
+    "register_fused_strategy",
     "get_normal_strategy",
     "get_low_latency_strategy",
+    "get_fused_strategy",
     # Normal strategies
     "DefaultNormalCommStrategy",
     "AlltoAllNormalCommStrategy",
     # Low latency strategies
     "DefaultLowLatencyCommStrategy",
     "OpsLowLatencyCommStrategy",
+    # Fused strategies
+    "DeepEPFusedStrategy",
+    "MegaMoeFusedStrategy",
 ]
